[!] fix issue #220, fix compile warning (#221)

+++ src/transport/xqc_packet_parser.c
@@ -992,12 +992,12 @@ xqc_packet_parse_retry(xqc_connection_t *c, xqc_packet_in_t *packet_in)
     }
 
     /* Token (..) */
-    uint8_t retry_token_len = XQC_BUFF_LEFT_SIZE(pos, end) - XQC_RETRY_INTEGRITY_TAG_LEN;
+    int retry_token_len = XQC_BUFF_LEFT_SIZE(pos, end) - XQC_RETRY_INTEGRITY_TAG_LEN;
     if (retry_token_len > XQC_MAX_TOKEN_LEN) {
         xqc_log(c->log, XQC_LOG_ERROR, "|retry token length exceed XQC_MAX_TOKEN_LEN|%d|", retry_token_len);
         return -XQC_EILLPKT;
     }
-    
+
     memcpy(c->conn_token, pos, retry_token_len);
     c->conn_token_len = retry_token_len;
